agent_guard/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-agent_guard"
version = "3.0.0"
description = "AgentGuard V3: Adversarial Permission Governance Simulator for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # OpenAI client for inference.py
    "openai>=1.0.0",
    # Gradio for interactive demo UI
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "agent_guard.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["agent_guard", "agent_guard.server"]
package-dir = { "agent_guard" = ".", "agent_guard.server" = "server" }